Refactor theme directory
